Amélioration des commentaires des classes metier, test et util.

+++ ProjetPOO/src/test/TestRecherche.java
@@ -1,8 +1,3 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 package test;
 
 import java.util.ArrayList;
@@ -11,15 +6,16 @@ import util.Reader;
 import util.Recherche;
 
 /**
- *
- * @author nicol
+ * Classe permettant de tester la classe Recherche.
+ * Lit le fichier d'instances, et génère une solution simple de parcours de la zone de picking.
  */
 public class TestRecherche {
-    public static void main(String[] args) throws Exception {    
-        //public Recherche(ArrayList<Order> orderList, ArrayList<Product> productList, int nbColisMax, int weightMax_parcel, int volumeMax_parcel)
+    public static void main(String[] args) throws Exception {
+        
         Reader r = new Reader("instance_simple.txt");
-        Recherche sol = new Recherche(r.getOrders(),r.getProducts(),r.getNbBoxesTrolley(),(int) r.getCapaBox().get(0), (int) r.getCapaBox().get(1));
+        Recherche sol = new Recherche( r.getOrders(), r.getProducts(), r.getNbBoxesTrolley(), r.getCapaBox().get(0), r.getCapaBox().get(1));
         ArrayList<Trolley> solutions = new ArrayList();
+        
         solutions = sol.lookup();
         System.out.println(solutions);
     }
